… encode_compile.go