Good news. In Unmarshal's benchmark, go-json is faster than gojay .

Faster 20% - 30% than master branch 

```
benchmark                                                     old ns/op     new ns/op     delta
Benchmark_Decode_SmallStruct_Unmarshal_EncodingJson-16        2547          2436          -4.36%
Benchmark_Decode_SmallStruct_Unmarshal_JsonIter-16            793           758           -4.41%
Benchmark_Decode_SmallStruct_Unmarshal_GoJay-16               461           444           -3.69%
Benchmark_Decode_SmallStruct_Unmarshal_GoJayUnsafe-16         409           405           -0.98%
Benchmark_Decode_SmallStruct_Unmarshal_SegmentioJson-16       1039          1024          -1.44%
Benchmark_Decode_SmallStruct_Unmarshal_GoJson-16              498           352           -29.32%
Benchmark_Decode_SmallStruct_Unmarshal_GoJsonNoEscape-16      516           344           -33.33%
Benchmark_Decode_SmallStruct_Stream_EncodingJson-16           2751          2738          -0.47%
Benchmark_Decode_SmallStruct_Stream_JsonIter-16               972           992           +2.06%
Benchmark_Decode_SmallStruct_Stream_GoJay-16                  562           555           -1.25%
Benchmark_Decode_SmallStruct_Stream_SegmentioJson-16          5293          5553          +4.91%
Benchmark_Decode_SmallStruct_Stream_GoJson-16                 987           783           -20.67%
Benchmark_Decode_MediumStruct_Unmarshal_EncodingJson-16       15134         15578         +2.93%
Benchmark_Decode_MediumStruct_Unmarshal_JsonIter-16           5382          5352          -0.56%
Benchmark_Decode_MediumStruct_Unmarshal_GoJay-16              3151          3178          +0.86%
Benchmark_Decode_MediumStruct_Unmarshal_GoJayUnsafe-16        2658          2721          +2.37%
Benchmark_Decode_MediumStruct_Unmarshal_SegmentioJson-16      5071          5366          +5.82%
Benchmark_Decode_MediumStruct_Unmarshal_GoJson-16             2833          2295          -18.99%
Benchmark_Decode_MediumStruct_Unmarshal_GoJsonNoEscape-16     2803          2372          -15.38%
Benchmark_Decode_MediumStruct_Stream_EncodingJson-16          18059         17934         -0.69%
Benchmark_Decode_MediumStruct_Stream_JsonIter-16              6334          6171          -2.57%
Benchmark_Decode_MediumStruct_Stream_GoJay-16                 4328          4273          -1.27%
Benchmark_Decode_MediumStruct_Stream_SegmentioJson-16         13549         13400         -1.10%
Benchmark_Decode_MediumStruct_Stream_GoJson-16                4670          4089          -12.44%
Benchmark_Decode_LargeStruct_Unmarshal_EncodingJson-16        197295        195806        -0.75%
Benchmark_Decode_LargeStruct_Unmarshal_JsonIter-16            97707         97652         -0.06%
Benchmark_Decode_LargeStruct_Unmarshal_GoJay-16               36738         35992         -2.03%
Benchmark_Decode_LargeStruct_Unmarshal_GoJayUnsafe-16         31821         31463         -1.13%
Benchmark_Decode_LargeStruct_Unmarshal_SegmentioJson-16       89198         89756         +0.63%
Benchmark_Decode_LargeStruct_Unmarshal_GoJson-16              39769         32051         -19.41%
Benchmark_Decode_LargeStruct_Unmarshal_GoJsonNoEscape-16      40516         31836         -21.42%
Benchmark_Decode_LargeStruct_Stream_EncodingJson-16           229008        215475        -5.91%
Benchmark_Decode_LargeStruct_Stream_JsonIter-16               113101        108454        -4.11%
Benchmark_Decode_LargeStruct_Stream_GoJay-16                  45679         43176         -5.48%
Benchmark_Decode_LargeStruct_Stream_SegmentioJson-16          151604        147518        -2.70%
Benchmark_Decode_LargeStruct_Stream_GoJson-16                 70537         56056         -20.53%
```